Greedy Stitcher split tracks again if they overlap
This was a large problem, that A crosses over B, then B is only visible but at the same location as A was before, then B takes over A's ID and once A is visible again, A gets a new ID (C).
Now, B does not take over the ID anymore and we have two new tracks after which the actual Tracker can hopefully merge better.